fix tooltip positioning when nav is hidden
